requirements.txt
# frappe -- https://github.com/frappe/frappe is installed via 'bench init'
pyqrcode==1.2.1
python-barcode==0.15.1